+++ scripts/diff/json.rs
@@ -2178,7 +2178,7 @@ impl ::rusty_lr::parser::Parser for JsonParser {
                         let term_class = JsonTerminalClasses::from_usize(
                             (val & 0x7fff) as usize,
                         );
-                        let state = ((val >> 15) & 0xffff) as u8;
+                        let state = ((val >> 15) & 0xffff) as usize;
                         let push = (val >> 31) != 0;
                         shift_goto_map_term
                             .push((
@@ -2196,7 +2196,7 @@ impl ::rusty_lr::parser::Parser for JsonParser {
                         let nonterm = JsonNonTerminals::from_usize(
                             (val & 0x7fff) as usize,
                         );
-                        let state = ((val >> 15) & 0xffff) as u8;
+                        let state = ((val >> 15) & 0xffff) as usize;
                         let push = (val >> 31) != 0;
                         shift_goto_map_nonterm
                             .push((
@@ -2216,7 +2216,7 @@ impl ::rusty_lr::parser::Parser for JsonParser {
                         let len = REDUCE_DATA[idx + 1] as usize;
                         let mut rules = Vec::with_capacity(len);
                         for r_idx in 0..len {
-                            rules.push(REDUCE_DATA[idx + 2 + r_idx] as u8);
+                            rules.push(REDUCE_DATA[idx + 2 + r_idx] as usize);
                         }
                         reduce_map.push((term_class, rules));
                         idx += 2 + len;
